fix: skip messaging via Slack, comment upon closing instead (#37)
