+++ .git-blame-ignore-revs
@@ -1,0 +1,2 @@
+# https://github.com/shopware/gh-project-automation/pull/30
+23772cea8e988928e3a364003eb4b2dfded7f4e5
