+++ src/services/issue.ts
@@ -9,18 +9,15 @@ import {
 
 import {
     addComment,
-    addLabelToLabelable, closePullRequest,
+    addLabelToLabelable,
     findIssueWithProjectItems,
     findPRWithProjectItems,
     getCommentsForIssue,
     getIssuesByProject,
     getLabelByName,
-    getPullRequests, getVerifiedDomainEmails,
+    getPullRequests,
     jiraHost
 } from "../api";
-
-import {sendSlackMessageForEMail} from "./slack";
-import {isDryRun} from "../util/dry_run";
 
 export const docIssueReference = Buffer.from("doc-issue-created").toString("base64");
 
@@ -276,54 +273,5 @@ export async function findWithProjectItems(toolkit: Toolkit) {
         return await findPRWithProjectItems(toolkit, toolkit.context.payload.pull_request.number);
     } else {
         throw new Error('only issue and pull_request events are supported');
-    }
-}
-
-/**
- * manageOldPullRequests checks for old pull requests and sends a reminder message to the assignee.
- *
- * @param toolkit - Octokit instance. See: https://octokit.github.io/rest.js
- * @param organization - The GitHub organization to check for old pull requests.
- * @param days - Consider pull requests old after this many days of inactivity.
- * @param close - If true, the pull request will be closed after sending the reminder.
- */
-export async function manageOldPullRequests(toolkit: Toolkit, organization: string = "shopware", days: number = 7, close: boolean = false) {
-    const pullRequests = await getPullRequests(
-        toolkit,
-        `org:${organization} is:pr is:open updated:<${new Date(Date.now() - days * 24 * 60 * 60 * 1000).toISOString()}`
-    );
-
-    for (const pr of pullRequests) {
-        const assignee = pr.assignees.nodes[0];
-
-        if (!assignee) {
-            toolkit.core.debug(`Pull request ${pr.repository.owner.login}/${pr.repository.name}#${pr.number} has no assignee, skipping.`);
-
-            continue;
-        }
-
-        const emails = await getVerifiedDomainEmails(toolkit, assignee.login, organization);
-
-        if (emails.length < 1) {
-            continue; // No verified domain emails found for the assignee, abort.
-        }
-
-        const baseMsg = `Hi @${assignee.login}, the pull request [${pr.repository.owner.login}/${pr.repository.name}#${pr.number}](${pr.url}) has not been updated in over ${days} days. Please take a look and update it if needed.`
-        const closeMsg = `The pull request has been closed automatically. If you would like to continue working on it, please feel free to re-open it!`;
-        const message = close ? `${baseMsg}\n${closeMsg}` : baseMsg;
-
-        if (close) {
-            if (isDryRun()) {
-                toolkit.core.info(`[DRY_RUN]\tWould close pull request ${pr.repository.owner.login}/${pr.repository.name}#${pr.number}.`);
-            } else {
-                await closePullRequest(toolkit, pr.id);
-            }
-        }
-
-        if (isDryRun()) {
-            toolkit.core.info(`[DRY_RUN]\tWould send message to @${assignee.login}.`);
-        } else {
-            await sendSlackMessageForEMail(toolkit, emails, message);
-        }
     }
 }
